2 small changes only - moved utils package to fix maven getting confused between plugins - made [TownyCultures] a soft depend so that Siegewar will load after it
