+++ src/main/java/com/gmail/goosius/siegewar/settings/CommentedConfiguration.java
@@ -8,7 +8,7 @@ import org.yaml.snakeyaml.DumperOptions;
 import org.yaml.snakeyaml.Yaml;
 import org.yaml.snakeyaml.representer.Representer;
 
-import io.github.townyadvanced.util.FileMgmt;
+import com.gmail.goosius.siegewar.utils.FileMgmt;
 
 import java.io.File;
 import java.io.IOException;
