+++ src/main/java/com/gmail/goosius/siegewar/settings/Translation.java
@@ -5,7 +5,7 @@ import com.palmergames.bukkit.towny.command.HelpMenu;
 import com.palmergames.bukkit.util.Colors;
 import com.palmergames.util.StringMgmt;
 
-import io.github.townyadvanced.util.FileMgmt;
+import com.gmail.goosius.siegewar.utils.FileMgmt;
 
 import org.bukkit.configuration.InvalidConfigurationException;
