+++ src/main/resources/plugin.yml
@@ -6,7 +6,7 @@ author: [Goosius,LlmDl]
 website: 'townyadvanced.github.io'
 prefix: ${project.artifactId}
 depend: [Towny]
-softdepend: [townycultures,dynmap]
+softdepend: [TownyCultures,dynmap]
 
 description: A war system made by Goosius for Towny.
